fix(driver-app): register FCM background handler at bundle entry
Ride offers are data-only FCM messages. When the app is killed, Android
delivers them via a headless JS launch where expo-router route modules
(app/_layout.tsx) never execute — so the background handler registered
there never existed and offers were silently dropped: no notification,
no sound, nothing on the lock screen.

Move setBackgroundMessageHandler + notifee.onBackgroundEvent into
services/backgroundMessaging.ts and register from index.js, the real
bundle entry that always runs in headless mode. Foreground listeners
stay in _layout.tsx. PENDING_OFFER_KEY/PENDING_ACTION_KEY now exported
from the new module.
