+++ driver-app/app/_layout.tsx
@@ -37,17 +37,16 @@ import {
   initFirebaseServices,
   requestNotificationPermission,
   requestPushPermissionAndGetToken,
-  setBackgroundMessageHandler,
   onTokenRefresh,
   getAppCheckToken,
 } from '@shared/services/firebase';
+import { PENDING_ACTION_KEY } from '../services/backgroundMessaging';
 import { setAppCheckTokenProvider } from '@shared/api/client';
 // Notifee — rich notifications (heads-up + full-screen intent + Accept/Decline
 // action buttons). Lazy-required because the native module isn't linked in
 // Expo Go / web; we treat the import failure the same as Expo Go: no-op.
 let notifee: any = null;
 let parseRideOfferEvent: any = null;
-let displayRideOfferNotification: any = null;
 let dismissRideOfferNotification: any = null;
 let ensureNotifeeReady: any = null;
 if (Platform.OS === 'android' || Platform.OS === 'ios') {
@@ -55,7 +54,6 @@ if (Platform.OS === 'android' || Platform.OS === 'ios') {
     notifee = require('@notifee/react-native').default;
     const svc = require('../services/notifeeService');
     parseRideOfferEvent = svc.parseRideOfferEvent;
-    displayRideOfferNotification = svc.displayRideOfferNotification;
     dismissRideOfferNotification = svc.dismissRideOfferNotification;
     ensureNotifeeReady = svc.ensureNotifeeReady;
   } catch (e) {
@@ -139,118 +137,13 @@ if (Notifications) {
   });
 }
 
-// 2. Background FCM handler. Must be registered at module top level,
-//    outside of any React component, so the JS runtime wakes when a
-//    message arrives while the app is backgrounded or killed.
-//    Persists the full ride-offer payload to AsyncStorage so the driver
-//    home screen can hydrate the offer panel instantly on cold start
-//    (useDriverDashboard.ts reads PENDING_OFFER_KEY on mount).
-const PENDING_OFFER_KEY = 'spinr_pending_ride_offer';
-const PENDING_ACTION_KEY = 'spinr_pending_notifee_action';
-setBackgroundMessageHandler(async (remoteMessage: any) => {
-  const data = remoteMessage?.data || {};
-  if (data?.type === 'new_ride_assignment' && data?.ride_id) {
-    // FCM data uses same keys as WS dispatch_payload, all stringified.
-    const safeParse = <T,>(s: any): T | undefined => {
-      if (!s || s === 'null' || s === 'None') return undefined;
-      if (typeof s !== 'string') return s as T;
-      try { return JSON.parse(s) as T; } catch { return undefined; }
-    };
-    const toNum = (v: any): number | undefined => {
-      if (!v || v === '' || v === 'None') return undefined;
-      const n = parseFloat(v);
-      return Number.isFinite(n) ? n : undefined;
-    };
-    const fare = toNum(data.fare) ?? 0;
-    const totalBonus = toNum(data.total_bonus) ?? 0;
-    const surgeMultiplier = toNum(data.surge_multiplier);
-    const incentives = safeParse<any[]>(data.incentives);
-    const questHint = safeParse<any>(data.quest_hint);
-
-    // 1. Persist the full offer payload so the in-app panel can hydrate
-    //    instantly on cold start (driver dashboard reads on mount).
-    try {
-      const AsyncStorage = require('@react-native-async-storage/async-storage').default;
-      await AsyncStorage.setItem(
-        PENDING_OFFER_KEY,
-        JSON.stringify({
-          ride_id: data.ride_id,
-          booking_id: data.booking_id || data.ride_id,
-          pickup_address: data.pickup_address || '',
-          dropoff_address: data.dropoff_address || '',
-          pickup_lat: toNum(data.pickup_lat) ?? 0,
-          pickup_lng: toNum(data.pickup_lng) ?? 0,
-          dropoff_lat: toNum(data.dropoff_lat) ?? 0,
-          dropoff_lng: toNum(data.dropoff_lng) ?? 0,
-          fare,
-          distance_km: toNum(data.distance_km),
-          duration_minutes: toNum(data.duration_minutes),
-          rider_name: data.rider_name || undefined,
-          rider_rating: toNum(data.rider_rating),
-          requires_wav: data.requires_wav === 'true' || data.requires_wav === 'True',
-          countdown_seconds: toNum(data.countdown_seconds),
-          offer_expires_at: data.offer_expires_at || undefined,
-          surge_multiplier: surgeMultiplier,
-          incentives,
-          total_bonus: totalBonus || undefined,
-          quest_hint: questHint,
-          payment_method: data.payment_method || undefined,
-        }),
-      );
-    } catch (e) {
-      console.warn('[Push] Failed to persist background ride offer:', e);
-    }
-
-    // 2. Surface the Uber-style heads-up + full-screen-intent notification
-    //    via Notifee. This is what the driver actually sees on the lock
-    //    screen, with Accept/Decline buttons.
-    if (displayRideOfferNotification) {
-      try {
-        await displayRideOfferNotification({
-          ride_id: data.ride_id,
-          booking_id: data.booking_id || data.ride_id,
-          pickup_address: data.pickup_address,
-          dropoff_address: data.dropoff_address,
-          fare,
-          total_bonus: totalBonus,
-          distance_km: toNum(data.distance_km),
-          duration_minutes: toNum(data.duration_minutes),
-          surge_multiplier: surgeMultiplier,
-          rider_name: data.rider_name,
-          incentives_count: Array.isArray(incentives) ? incentives.length : 0,
-          countdown_seconds: toNum(data.countdown_seconds),
-          offer_expires_at: data.offer_expires_at || undefined,
-        });
-      } catch (e) {
-        console.warn('[Notifee] displayRideOfferNotification failed:', e);
-      }
-    }
-  }
-});
-
-// Notifee background event listener — fires when the user taps Accept/Decline
-// from the lock screen or notification shade while the app is killed. Must be
-// registered at module top level (mirrors the FCM background handler). We
-// stash the action in AsyncStorage so useDriverDashboard.ts can consume it on
-// mount and call accept/decline against the backend.
-if (notifee && parseRideOfferEvent) {
-  notifee.onBackgroundEvent(async (event: any) => {
-    const parsed = parseRideOfferEvent(event);
-    if (!parsed || !parsed.ride_id) return;
-    try {
-      const AsyncStorage = require('@react-native-async-storage/async-storage').default;
-      await AsyncStorage.setItem(
-        PENDING_ACTION_KEY,
-        JSON.stringify({ action: parsed.action, ride_id: parsed.ride_id, ts: Date.now() }),
-      );
-      // Dismiss the notification so the driver doesn't see stale "Accept"
-      // buttons while we open the app.
-      if (dismissRideOfferNotification) await dismissRideOfferNotification();
-    } catch (e) {
-      console.warn('[Notifee] background action persist failed:', e);
-    }
-  });
-}
+// 2. Background FCM + Notifee killed-state handlers are registered in
+//    index.js via services/backgroundMessaging.ts — they MUST live at the
+//    real bundle entry. When the app is killed, Android delivers data-only
+//    ride offers through a headless JS launch where route modules (this
+//    file) never execute, so a handler registered here would never exist
+//    and the offer would be silently dropped. PENDING_ACTION_KEY is shared
+//    with the foreground listeners below.
 
 // DV-9: Route push-notification taps to the correct in-app screen.
 // new_ride_assignment → driver home (offer panel hydrates from AsyncStorage);
